movie_cast/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==5.0.0
Markdown==3.4.1
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
zipp==3.10.0
